* Modified Color banding frequency to period: +colour spectrum will repeat for each period. All clouds will use same colour scheme -High max and min not used any more +sine changed to float and sine of 120 and 240 deg converted to radians with higher accuracy +parameter frequency from int type to double +updated UI dialog to double +command line call frequency changed to double

+++ libs/qCC_db/ccPointCloud.h
@@ -541,7 +541,7 @@ public: //other methods
 		\param freq banding frequency
 		\return success
 	**/
-	bool setRGBColorByBanding(unsigned char dim, int freq);
+	bool setRGBColorByBanding(unsigned char dim, double freq);
 
 	//! Sets RGB colors with current scalar field (values & parameters)
 	/** \return success
